refactor e2e tests
